+++ paddle/phi/kernels/cpu/pool_grad_kernel.cc
@@ -64,7 +64,7 @@ PD_REGISTER_KERNEL(fractional_max_pool2d_grad,
                    float,
                    double,
                    phi::float16) {
-  kernel->InputAt(1).SetDataType(phi::CppTypeToDataType<int>::Type());
+  kernel->InputAt(1).SetDataType(phi::CppTypeToDataType<int64_t>::Type());
 }
 
 PD_REGISTER_KERNEL(fractional_max_pool3d_grad,
@@ -74,5 +74,5 @@ PD_REGISTER_KERNEL(fractional_max_pool3d_grad,
                    float,
                    double,
                    phi::float16) {
-  kernel->InputAt(1).SetDataType(phi::CppTypeToDataType<int>::Type());
+  kernel->InputAt(1).SetDataType(phi::CppTypeToDataType<int64_t>::Type());
 }
